add shopping cart

+++ lib/common/utils/utils.dart
@@ -1,4 +1,4 @@
-String priceFormat(int price) {
+String priceFormat(double price) {
   return price.toStringAsFixed(2).replaceAllMapped(
         RegExp(r'(\d{1,3})(?=(\d{3})+(?!\d))'),
         (Match m) => '${m[1]},',
